requirements.txt
Pillow==10.3.0
Requests==2.32.3
setuptools==70.0.0
streamlit==1.35.0
